// src/race/events/race-event.enum.ts
export enum RaceEvents {
  CONNECTED = 'race:connected',
  JOIN_RACE = 'race:join',
  LEAVE_RACE = 'race:leave',
  START_COUNTDOWN = 'race:start-countdown',
  TYPING_PROGRESS = 'race:typing-progress',
  EXTEND_START_TIME = 'race:extend-start-time',
  END_RACE = 'race:end',
  BROADCAST_RESULTS = 'race:broadcast-results',
  UPDATE_PARTICIPANTS = 'race:update-participants',
  START_RACE = 'race:start-race',
}